feat: Extract AppConfig types and AppLoggerKeys to domain module (Phase 23)
Move pure Kotlin config types (AppConfig, Server, InitialData,
FetchOnViewModelInit) and AppLoggerKeys object from androidApp's
config package to domain module. APP_CONFIG singleton stays in
androidApp since it reads BuildConfig.
